refactor(tree): move initializeForest to feature-libraries
