+++ packages/dds/tree/src/shared-tree/independentView.ts
@@ -14,7 +14,6 @@ import {
 	type RevisionTag,
 	RevisionTagCodec,
 	TreeStoredSchemaRepository,
-	initializeForest,
 	type ITreeCursorSynchronous,
 	mapCursorField,
 } from "../core/index.js";
@@ -27,6 +26,7 @@ import {
 	chunkTree,
 	defaultChunkPolicy,
 	TreeCompressionStrategy,
+	initializeForest,
 } from "../feature-libraries/index.js";
 // eslint-disable-next-line import/no-internal-modules
 import type { Format } from "../feature-libraries/schema-index/format.js";
