+++ packages/dds/tree/src/test/domains/json/jsonCursor.bench.ts
@@ -13,7 +13,6 @@ import {
 	type ITreeCursor,
 	type JsonableTree,
 	TreeStoredSchemaRepository,
-	initializeForest,
 	moveToDetachedField,
 } from "../../../core/index.js";
 import {
@@ -28,6 +27,7 @@ import {
 	cursorForJsonableTreeNode,
 	cursorForMapTreeNode,
 	defaultSchemaPolicy,
+	initializeForest,
 	jsonableTreeFromCursor,
 	mapTreeFromCursor,
 } from "../../../feature-libraries/index.js";
